BUG: stats: fix wrong shape output of _stats method for burr and fisk distributions (#13238)
As described in gh-13234, 'burr' and 'fisk' distributions returned a bad shaped array when scalar arguments were passed to the 'moment' method leading to an API inconsistency. This was because the call to '_stats' method from the 'moment' method when n<5 and n>1 returned a bad shaped array. The '_stats' method (in 'burr' distribution) has been fixed to return floating points when scalar arguments are passed. Regression tests for it have also been added.

Closes gh-13234

Co-authored-by: Warren Weckesser <warren.weckesser@gmail.com>