MAINT: add tests of trivial cost matrices for linear sum assignment
Similar tests used to be part of the general test that the method finds
the expected results, but as pointed out by @rgommers in
https://github.com/scipy/scipy/pull/13175, those tests disappeared
during an earlier refactoring, which aimed to reduce duplication between
these tests and those of `min_weight_full_bipartite_matching`. Here, we
reintroduce the tests that went missing.