BUG: stats: fix the _cdf method of rv_discrete class
As described in gh-13280, the cdf method failed for 'hypergeom' and 'nhypergeom' distributions when invalid arguments were passed. The fix is to fill in the 'nan' values for invalid arguments after filling in the '1' values for quantiles greater than the upper bound in the 'cdf' method of 'rv_discrete' class. This fix has been implemented and regression tests for it have been added.