Remove NormalQMC, add MultinomialQMC
Remmoves the NormalQMC engine and instead makes the `cov` argument to `MultivariateNormalQMC` optional and interprets an omitted as the identity covariance matrix.
Adds a `covar_root` arg to the `MultivariateNormalQMC` constructor that allows passing in a root decomposition of the covariance matrix so that factorization does not need to be performed if already available.
Removes the `multinomial_qmc` function in favor of a `MultinomialQMC` engine.