BUG: stats: Don't raise ZeroDivisionError in _unequal_var_ttest_denom
When both standard deviations given to ttest_ind_from_stats were Python
floating point zeros and the parameter equal_var was False, the private
function _unequal_var_ttest_denom would raise a ZeroDivisionError.  The
code that raised the exception is wrapped in a np.errstate context
manager to catch such errors, but that context manager does not catch
an exception if it is raised by the division of Python floats.

The fix is to convert the inputs of ttest_ind_from_stats to NumPy arrays
before doing any calculations.  Then the exception raised by the
calculation in _unequal_var_ttest_denom will be caught by the context
manager.

Closes gh-6409.