DOC: minor formatting.
The missing colon make the whole block appear as comment for sphinx

Us the occasion to fix the see-also/parameter in the same file; the missing colon
(and when there are no space before) make numpydoc parse the section
sometime incorrectly. While this can be hard or invisible in current
scipy/numpy theme; this is semantically incorrect.