TST: fix failures due to new FutureWarnings in NumPy 1.21.dev0
The warnings were:

FutureWarning: Promotion of numbers and bools to strings is deprecated.

The `test_lfilter_bad_object` changes from ValueError to TypeError.
The changes in behaviour should be due to the refactoring in
the NumPy dtype system. There's not much point keeping these
tests, more trouble than they're worth.

Closes gh-13472