BUG: sparse: Create a utility function `getdata`
The utility function `getdata` is a thin wrapper of
`np.array(data, copy=copy, dtype=dtype)`; it also calls
`getdtype` to check the dtype of the data array.  A
warning is generated if the dtype is object.

The new function is used in coo_matrix and bsr_matrix
when the input to the constructor has the form (data (i, j)).

Note that this removes the default dtype of float for
BSR matrices.  Now if `data` is given as, say, a Python
list of lists of integers, and an explicit dtype is not
given, the sparse matrix will be an integer type.  This
is consistent with how csr_matrix and csc_matrix work.

Closes gh-13585.