MAINT, TST: stats: centralize invalid parameters list for all the distributions (#13507)
* MAINT, TST: stats: add invalid parameters list for all the distributions

A list of invalid parameters for all the distribution was added in gh-13295. The list of invalid parameters for discrete distributions was taken from gh-13280. This list has now been centralized in the '_dist_params.py' file where invalid parametrization of new distributions can easily be added and those tests can incorporate them from this central list.

* TST: stats: add test to match valid and invalid parameter lists

A test has been added to check if the valid parameter list and invalid parameter list match each other. A comment explaining why some distributions are skipped in test_support has been added.

* TST: stats: reomve the try except block in cdf test

The try except block checked for a non-string distribution
in the `invdistdiscrete` list. But this is never the case
currently, so the block was un-necessary. It has been removed
in this commit.

* TST: stats: remove try except and condition from lists

Removed a try except block as per the reason mentioned in the previous commit and remove the condition checking for non-string distributions from all the lists (in 'test_distr_params_lists') other than 'distdiscrete' where it is needed.