ENH: stats: add 'alternative' keyword to some normality tests. (#13549)
* ENH: stats: add 'alternative' keyword to some normality tests.

This commit adds `alternative` parameter to `skewtest`, `kurtosistest`,
and `ranksums` functions to allow one-tailed hypothesis testing. An example
elucidating the use of this parameter has been added in the documentation
suite of each of the changed functions. Necessary tests to ensure full
coverage have also been added.

A bit unrelated, but necessary change in the tests of the `ranksums`
function has also been made. The tests of `ranksums` did not validate
the outputs previously. Hence, a more comprehensive testsuite, comparing
the outputs with R's `wilcox.test`, has been added.

Other normality tests (present in the table gh-12506), namely `kendalltau`,
`ansari`, and `mood`, are a bit tricky to handle. Hence, I have left them
unchanged for now.

* DOC: stats: fix the varying output in ranksums.

* TST: stats: use sk_skew and sk_kurt in alternative tests

As pointed out in the review, using 'sk_skew' and 'sk_kurt' instead of 'sk_skew_{less/greater}' and 'sk_kurt_{less/greater}' in the tests for 'alternative' keyword should not be very confusing and helps remove unnecessary assignments.

* TST: stats: test ranksums using mannwhitneyu statistic

R's `wilcox.test` used ties correction which gave a p-value
that did not entirely match `ranksums`. This has been fixed by first
calculating the `mannwhitneyu` statistic and then calculating the
p-value in the test itself, hence, circumventing ties correction.

* DOC: stats: add more informative docs for alternative

As discussed with @mdhaber, this commit adds more informative
docs for the `alternative` parameter of the `skewtest`, `kurtosistest`,
and `ranksums` functions.

* DOC: stats: fix the doc string of the alternative parameter

The docs of the alternative parameter were not rendered properly previously. This has been fixed now.

* DOC, TST: stats: add more intuitive docs and tests

As suggested by @mdhaber, this commit adds more intuitive tests for the 'skewtest' and 'kurtosistest'. Also, the documentation of the 'alternative' parameter has been changed so that the user need not be familiar with theoretical details of the tests.

* TST: stats: simplify TestRankSums

* DOC: stats: adjust description of alternatives, remove inadvertent whitespace

* DOC: stats: remove some potentially confusing information about kurtosistest

Co-authored-by: Matt Haberland <mhaberla@calpoly.edu>