BUG: stats: Override _fitstart for the invweibull distribution. (#13712)
* BUG: stats: Override _fitstart for the invweibull distribution.

When no initial guess is given, the _fitstart method uses 1.0
for the initial guess of shape parameters, and then uses the
distribution's moment to create a better guess.  The invweibull
distribution doesn't have a first moment when the shape parameter
c is 1, so _fitstart is overridden to use 2.0.

This fixes a warning that occurs when calling invweibull.fit.  E.g.

    >>> invweibull.fit([1, 2, 3, 10, 20], floc=0)
    [...]/scipy/stats/_distn_infrastructure.py:2494: RuntimeWarning:
    invalid value encountered in double_scalars
      Lhat = muhat - Shat*mu
    (1.094936085143949, 0, 2.435695425443909)

After this PR:

    >>> invweibull.fit([1, 2, 3, 10, 20], floc=0)
    (1.0949396631669803, 0, 2.435663730822117)

Co-authored-by: Matt Haberland <mhaberla@calpoly.edu>